moved geojson Parser
